+++ students/myungjun/project_westagram/asgi.py
@@ -1,16 +1,16 @@
 """
-ASGI config for westagram project.
+ASGI config for project_westagram project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/3.2/howto/deployment/asgi/
+https://docs.djangoproject.com/en/dev/howto/deployment/asgi/
 """
 
 import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'westagram.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'project_westagram.settings')
 
 application = get_asgi_application()
