--- students/wonseok/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/home/wslee/miniconda3/envs/westargram/bin/python"
-}
