--- students/wonseok/project_westagram/urls.py
@@ -1,5 +1,0 @@
-# from django.urls import path, include
-
-urlpatterns = [
-    # path("user", include("user.urls"))
-]
